Serveur_CNN/requirements.txt
scipy
scikit-learn
keras
absl-py==0.6.1
astor==0.7.1
Bashutils==0.0.4
Django==2.1.2
djangorestframework==3.9.0
gast==0.2.0
grpcio==1.17.1
h5py==2.9.0
jsonfield==2.0.2
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
layers==0.1.5
Markdown==3.0.1
numpy==1.15.4
Pillow==5.3.0
protobuf==3.6.1
pytz==2018.6
PyYAML==3.13
sequential==1.0.0
six==1.12.0
tensorboard==1.12.1
tensorflow==1.12.0
termcolor==1.1.0
Werkzeug==0.14.1
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16